+++ src/commands/enrollRewardProgram.ts
@@ -19,17 +19,17 @@ import { BootstrapService, BootstrapUtils } from '../service';
 import { AnnounceService } from '../service/AnnounceService';
 import { CommandUtils } from '../service/CommandUtils';
 
-export default class EnrolRewardProgram extends Command {
-    static description = `It enrols the nodes in the rewards program by announcing the enrol transaction to the registration address.  You can also use this command to update the program registration when you change the node public key or server host.
+export default class EnrollRewardProgram extends Command {
+    static description = `It enrols the nodes in the rewards program by announcing the enroll transaction to the registration address.  You can also use this command to update the program registration when you change the node public key or server host.
 
 Currently, the only program that can be enrolled post-launch is 'SuperNode'.`;
 
     static examples = [
-        `$ symbol-bootstrap enrolRewardProgram`,
-        `$ symbol-bootstrap enrolRewardProgram --noPassword`,
-        `$ symbol-bootstrap enrolRewardProgram --useKnownRestGateways`,
-        `$ symbol-bootstrap enrolRewardProgram --password 1234 --url http://external-rest:3000`,
-        `$ echo "$MY_ENV_VAR_PASSWORD" | symbol-bootstrap enrolRewardProgram --url http://external-rest:3000`,
+        `$ symbol-bootstrap enrollRewardProgram`,
+        `$ symbol-bootstrap enrollRewardProgram --noPassword`,
+        `$ symbol-bootstrap enrollRewardProgram --useKnownRestGateways`,
+        `$ symbol-bootstrap enrollRewardProgram --password 1234 --url http://external-rest:3000`,
+        `$ echo "$MY_ENV_VAR_PASSWORD" | symbol-bootstrap enrollRewardProgram --url http://external-rest:3000`,
     ];
 
     static flags = {
@@ -39,7 +39,7 @@ Currently, the only program that can be enrolled post-launch is 'SuperNode'.`;
     };
 
     public async run(): Promise<void> {
-        const { flags } = this.parse(EnrolRewardProgram);
+        const { flags } = this.parse(EnrollRewardProgram);
         BootstrapUtils.showBanner();
         flags.password = await CommandUtils.resolvePassword(
             flags.password,
@@ -47,6 +47,6 @@ Currently, the only program that can be enrolled post-launch is 'SuperNode'.`;
             CommandUtils.passwordPromptDefaultMessage,
             true,
         );
-        return new BootstrapService(this.config.root).enrolRewardProgram(flags);
+        return new BootstrapService(this.config.root).enrollRewardProgram(flags);
     }
 }