+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 -   Fixed case issue validating keys when creating certificates.
 -   Updated Wallet to latest 1.0.1 release.
 -   Dropped NodeJS 10 support. Added Node LTS and Stable Support. 
+-   Renamed `enrol` to `enroll` for the Rewards Program.
 
 ## [1.0.3] - Mar-31-2021